% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epismthr.R
\name{epi_forecast}
\alias{epi_forecast}
\title{Creates an epi forecast based on an input tibble.}
\usage{
epi_forecast(data, look_ahead = 7, level = 95)
}
\arguments{
\item{data}{is a tibble that contains two columns: Date (date) and Cases (numeric)}

\item{look_ahead}{(default is 7), which is the forecast horizon}

\item{level}{(default is 95), which is confidence interval for the results}
}
\value{
A list containing plots, projections and all generated fable tibbles
}
\description{
\code{epi_forecast} returns a short term forecast of epi-cases
}
